+++ pkg/goatfile/parser.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 
 	"github.com/studio-b12/goat/pkg/errs"
-	"github.com/zekrotja/rogu/log"
 )
 
 // Parser parses a Goatfile.
@@ -406,16 +405,6 @@ func (t *Parser) parseBlock(req *requestParseChecker) error {
 		req.QueryParams = data
 
 	case optionNameHeader:
-		err := t.parseHeaders(req.Header)
-		if err != nil {
-			return err
-		}
-	case optionNameHeaders:
-		// TODO: Remove due to deprecation with release 1.0
-		log.Warn().Tag("Parser").
-			Field("pos", fmt.Sprintf("%d:%d", t.s.line, t.s.linepos)).
-			Msg("Option name [headers] is deprecated! Please use [header] instead! See the release notes of v0.8.0 for more information: " +
-				"https://github.com/studio-b12/goat/releases/tag/v0.8.0")
 		err := t.parseHeaders(req.Header)
 		if err != nil {
 			return err